requirements.txt
openai>=1.68.0
streamlit>=1.42.0
python-docx>=1.1.2
pypandoc>=1.15
python-dotenv>=1.0.1
pytest>=8.4.2